Move formatLoansAmount to utils/loans Add formatLoansAmount to PendingOffersTable
